Various bug fixes including rigid body change on player
1) Multiplayer player rigid body physics changed to continuous dynamic
to see if this helps fix the bug where the player goes through the
floor.
2) Level-removed some zombies to help fps and make the wave easier,
fixed the zombies spawning below ground, fixed wave 18 not triggering
because the trigger was during the zipline and the rigid body is
kinetic during that time.
3) Teleporter-if the player is already being teleported, ignore the
trigger event. This should fix the player becoming invisible bug.
4) PlayerControl-When attaching to the zipline, reset isChangingLanes
to false.